Move C2S functions to c2s.go
